Fix ShellCheck issues in cleanup scripts

+++ jjb/integration/include-raw-integration-cleanup-workspace.sh
@@ -1,9 +1,12 @@
+#!/bin/bash
 echo "Cleaning up the workspace..."
 
 # Leftover files from previous runs could be wrongly copied as results.
 # Keep the cloned integration/test repository!
-for file_or_dir in `ls -A -1 -I "test"`
+for file_or_dir in *
 # FIXME: Make this compatible with multipatch and other possible build&run jobs.
 do
-  rm -vrf "$file_or_dir"
+    if [ "$file_or_dir" != "test" ]; then
+        rm -vrf "$file_or_dir"
+    fi
 done
